requirements/requirements.txt
fastapi==0.100.1
uvicorn==0.23.1
pydantic==2.1.1
SQLAlchemy==2.0.19
async-exit-stack==1.0.1
async-generator==1.10
psycopg2-binary==2.9.6
aioredis==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiocache[redis]==0.12.1
python-dotenv==1.0.0
aredis==1.1.8
python-multipart==0.0.6